Fix double-delete in spike_device
This patch fixes a double delete crash observed in Spike device implementations. The root cause is that the sim emulator object might be deleted in run() and the destructor later tries to remove it again, resulting in SIGSEGV. This change removes the deletion of the sim in the device's destructor, leaving only the delete function in run().